test.py: pytest c++ facades should respect saving logs on success
BostFacade and UnitFacade saving the logs only when test failed,
ignoring the -s parameter that should allow save logs on success. This
PR adding checking this parameter.

Closes scylladb/scylladb#24596

+++ test/pylib/cpp/boost/boost_facade.py
@@ -145,8 +145,9 @@ class BoostTestFacade(CppTestFacade):
             )
             return [failure], ''
 
-        log_xml.unlink(missing_ok=True)
-        stdout_file_path.unlink(missing_ok=True)
+        if not self.save_log_on_success:
+            log_xml.unlink(missing_ok=True)
+            stdout_file_path.unlink(missing_ok=True)
 
         if results:
             return results, ''
